// api/src/jobs/talent/types.ts
export interface TalentJob {
  referencenumber: string;
  title: string;
  description: string;
  company: string;
  city: string;
  state: string;
  country: string;
  dateposted: string;
  url: string;
  expirationdate?: string;
  streetaddress?: string;
  postalcode?: string;
  jobtype?: string;
  isremote?: string;
  benefit?: string;
  category?: string;
  logo?: string;
  experience?: string;
  cpc?: string;
  salary?: string;
}